+++ RealmSwift/RealmConfiguration.swift
@@ -184,7 +184,19 @@ extension Realm {
 
          - note: Setting this property to `true` doesn't disable file format migrations.
          */
-        public var deleteRealmIfMigrationNeeded: Bool = false
+        public var deleteRealmIfMigrationNeeded: Bool {
+            set(newValue) {
+                if newValue && syncConfiguration != nil {
+                    throwRealmException("Cannot set 'deleteRealmIfMigrationNeeded' when sync is enabled ('syncConfig' is set).")
+                }
+                _deleteRealmIfMigrationNeeded = newValue
+            }
+            get {
+                return _deleteRealmIfMigrationNeeded
+            }
+        }
+
+        private var _deleteRealmIfMigrationNeeded: Bool = false
 
         /**
          A block called when opening a Realm for the first time during the