+++ CHANGELOG.md
@@ -4,7 +4,7 @@ x.y.z Release notes (yyyy-MM-dd)
 * None.
 
 ### Fixed
-* Not allow set 'deleteRealmIfMigrationNeeded' when sync is enabled.
+* Disallow `RLMRealmConfiguration.deleteRealmIfMigrationNeeded`/`Realm.Config.deleteRealmIfMigrationNeeded` when sync is enabled.
 
 <!-- ### Breaking Changes - ONLY INCLUDE FOR NEW MAJOR version -->